Moving LanguageModel class
