+++ src/python/PythonSDK/foundationallm/models/language_models/__init__.py
@@ -1,3 +1,4 @@
 """Language model models module"""
 from .language_model_provider import LanguageModelProvider
 from .language_model_type import LanguageModelType
+from .language_model import LanguageModel
