+++ src/python/PythonSDK/foundationallm/models/metadata/__init__.py
@@ -1,4 +1,3 @@
 """FoundationaLLM metadata models module"""
 from .agent import Agent
 from .data_source import DataSource
-from .language_model import LanguageModel
